feat: support OR search parameters (comma)
